+++ src/main/java/com/project/comgle/repository/MemberRepository.java
@@ -12,4 +12,6 @@ public interface MemberRepository extends JpaRepository<Member,Long> {
     Optional<Member> findByEmail(String email);
     Optional<Member> findByEmailAndCompany(String email,Company company);
     Optional<Member> findByMemberNameAndCompany(String memberName,Company company);
+    List<Member> findAllByCompany(Company company);
+
 }
