+++ src/main/java/com/project/comgle/service/AdminService.java
@@ -1,11 +1,15 @@
 package com.project.comgle.service;
 
+import com.project.comgle.dto.request.SignupRequestDto;
 import com.project.comgle.dto.response.MessageResponseDto;
+import com.project.comgle.entity.Company;
 import com.project.comgle.entity.Member;
 import com.project.comgle.entity.enumSet.PositionEnum;
 import com.project.comgle.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,6 +20,7 @@ import java.util.Optional;
 public class AdminService {
 
     private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
 
     // 회원 직책 수정 기능
     @Transactional
@@ -42,5 +47,49 @@ public class AdminService {
         findMember.get().updatePosition(position);
 
         return MessageResponseDto.of(HttpStatus.OK.value(), "변경 완료");
+    }
+
+    @Transactional
+    public ResponseEntity<MessageResponseDto> signup(SignupRequestDto signupRequestDto, Member member){
+
+        String password = passwordEncoder.encode(signupRequestDto.getPassword());
+        PositionEnum position = PositionEnum.valueOf(signupRequestDto.getPosition().trim().toUpperCase());
+
+        Company company = member.getCompany();
+
+        if(company == null){
+            throw new IllegalArgumentException("존재하지 않는 회사입니다.");
+        } else if ( member.getPosition() != PositionEnum.ADMIN) {
+            throw new IllegalArgumentException("ADMIN 권한이 필요합니다");
+        }
+
+        checkName(signupRequestDto.getMemberName(),company);
+        checkEmail(signupRequestDto.getEmail(),company);
+
+        Member newMember = Member.of(signupRequestDto,password,position,company);
+        memberRepository.save(newMember);
+        return ResponseEntity.ok(MessageResponseDto.of(HttpStatus.OK.value(), "회원가입 성공"));
+    }
+
+    @Transactional(readOnly = true)
+    public void checkEmail(String email,Company company){
+
+        Optional<Member> findMember = memberRepository.findByEmailAndCompany(email, company);
+
+        if(findMember.isPresent()){
+            throw new IllegalArgumentException("중복된 이메일이 존재합니다.");
+        }
+
+    }
+
+    @Transactional(readOnly = true)
+    public void checkName(String memberName,Company company) {
+
+        Optional<Member> findMember = memberRepository.findByMemberNameAndCompany(memberName, company);
+
+        if(findMember.isPresent()){
+            throw new IllegalArgumentException("중복된 사용자명가 존재합니다.");
+        }
+
     }
 }
